Update config.json
Set ?wanted to false for the quickpaylink. Making it the same as the normal paylink. 

This adds 1 extra screen to the checkout but gives context for some of the text fields and info on the checkout. I think it makes things less confusing. Although one more click which might be annoying to user. Not sure.